Somewhat polished.
