+++ src/view/Home.js
@@ -1,14 +1,13 @@
 
 import React from 'react'
 
-import {routeAbout} from 'actions'
-
 import {Section} from 'view/Shared/Structural'
 import {H1, P} from 'view/Shared/Typography'
+import {Link} from 'view/Shared/Link'
 
 export const Home = () => (
   <Section>
     <H1>Home</H1>
-    {/* <P>Head to the <Link to={routeAbout()}>About Page</Link>.</P> */}
+    <P>Head to the <Link href='/about'>About Page</Link>.</P>
   </Section>
 )
